requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.16.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.3.2
cryptography==42.0.8
distro==1.9.0
Django==4.1
django-widget-tweaks==1.5.0
exceptiongroup==1.2.0
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
joblib==1.3.0
msal==1.28.1
msal-extensions==1.1.0
multidict==6.0.5
numpy==1.23.5
openai==1.34.0
packaging==23.2
pandas==2.0.3
patsy==0.5.6
plotly==5.22.0
plotly-express==0.4.1
pluggy==1.3.0
portalocker==2.8.2
psycopg==3.1.14
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pyodbc==4.0.39
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
requests==2.32.3
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
statsmodels==0.14.2
tenacity==8.4.1
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4